requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
asyncio==3.4.3
redis==5.0.1
redis[hiredis]==5.0.1
aioredis==2.0.1
aiohttp==3.9.1
uvloop==0.19.0
hiredis==2.2.3